Photosynthesis

Photosynthesis model

Photosynthesis.leaf_photosynthesis! — Function

Per refactored Photosynthesis module, the only things one need to know is the public function leaf_photosynthesis! and some construtors from ClimaCache. See the examples in the methods below for details about how to use the function. The steps for computing photosynthetic rates are

Update temperature dependent variables using photosystem_temperature_dependence!

Calculate electron transport rate using photosystem_electron_transport!

Calculate RubisCO limited rate using rubisco_limited_rate!

Calculate light limited rate using light_limited_rate!

Calculate product limited rate using product_limited_rate!

Calculate gross and net rates using colimit_photosynthesis!

Update fluorescence related variables using photosystem_coefficients!

Supported methods are

leaf_photosynthesis!(leaf, air, mode)
leaf_photosynthesis!(leaf, air, mode, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/model.jl:65.

leaf_photosynthesis!(leaf, air, mode)
leaf_photosynthesis!(leaf, air, mode, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/model.jl:123.

Photosynthesis.leaf_photosynthesis! — Method

leaf_photosynthesis!(leaf::Leaf{FT}, air::AirLayer{FT}, mode::PCO₂Mode, p_i::FT = leaf.p_CO₂_i) where {FT<:AbstractFloat}

Updates leaf photosynthetic rates based on CO₂ partial pressure, given

leaf Leaf type structure that stores biophysical, reaction center, and photosynthesis model structures

air AirLayer structure for environmental conditions like O₂ partial pressure

mode PCO₂Mode that uses CO₂ partial pressure to compute photosynthetic rates

p_i Internal CO₂ partial pressure in Pa, default is leaf.p_CO₂_i

Examples

leaf = Leaf{Float64}("C3");
air  = AirLayer{Float64}();
mode = PCO₂Mode();
leaf_photosynthesis!(leaf, air, mode);
leaf_photosynthesis!(leaf, air, mode, 30.0);

Photosynthesis.leaf_photosynthesis! — Method

leaf_photosynthesis!(leaf::Leaf{FT}, air::AirLayer{FT}, mode::GCO₂Mode, g_lc::FT = leaf.g_CO₂) where {FT<:AbstractFloat}

Updates leaf photosynthetic rates based on CO₂ diffusive conductance, given

leaf Leaf type structure that stores biophysical, reaction center, and photosynthesis model structures

air AirLayer structure for environmental conditions like O₂ partial pressure

mode GCO₂Mode that uses CO₂ partial pressure to compute photosynthetic rates

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹], default is leaf.g_CO₂

Note that CO₂ partial pressures at leaf surface (stomatal opening) and in leaf internal airspace are updated.

Examples

leaf = Leaf{Float64}("C3");
air  = AirLayer{Float64}();
mode = GCO₂Mode();
leaf_photosynthesis!(leaf, air, mode);
leaf_photosynthesis!(leaf, air, mode, 0.1);

Temperature dependency

Photosynthesis.photosystem_temperature_dependence! — Function

This function update all the temperature dependencies of variables in a photosystem, such as Vcmax and Jmax. Supported methods are

photosystem_temperature_dependence!(psm, air, t)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/temperature.jl:150.

photosystem_temperature_dependence!(psm, air, t)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/temperature.jl:181.

photosystem_temperature_dependence!(psm, air, t)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/temperature.jl:211.

Photosynthesis.photosystem_temperature_dependence! — Method

photosystem_temperature_dependence!(psm::C3VJPModel{FT}, air::AirLayer{FT}, t::FT) where {FT<:AbstractFloat}

Update the temperature dependencies of C3 photosynthesis model, given

psm C3VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure

t Target temperature in K

Photosynthesis.photosystem_temperature_dependence! — Method

photosystem_temperature_dependence!(psm::C3CytochromeModel{FT}, air::AirLayer{FT}, t::FT) where {FT<:AbstractFloat}

Update the temperature dependencies of C3 photosynthesis model, given

psm C3CytochromeModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure

t Target temperature in K

Photosynthesis.photosystem_temperature_dependence! — Method

photosystem_temperature_dependence!(psm::C4VJPModel{FT}, air::AirLayer{FT}, t::FT) where {FT<:AbstractFloat}

Update the temperature dependencies of C3 photosynthesis model, given

psm C4VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure

t Target temperature in K

Photosynthesis.temperature_correction — Function

This function calculates the temperature correction ratio relative to the reference temperature. Supported methods are

temperature_correction(td, t; t_ref)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/temperature.jl:36.

temperature_correction(td, t; t_ref)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/temperature.jl:59.

temperature_correction(td, t; t_ref)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/temperature.jl:89.

Photosynthesis.temperature_correction — Method

temperature_correction(td::Arrhenius{FT}, t::FT; t_ref::FT = td.T_REF) where {FT<:AbstractFloat}

Return the correction ratio for a temperature dependent variable, given

td Arrhenius type temperature dependency struture

t Target temperature in K

t_ref Reference temperature in K, default is td.T_REF (298.15 K)

Photosynthesis.temperature_correction — Method

temperature_correction(td::ArrheniusPeak{FT}, t::FT; t_ref::FT = td.T_REF) where {FT<:AbstractFloat}

Return the correction ratio for a temperature dependent variable, given

td ArrheniusPeak type temperature dependency struture

t Target temperature in K

t_ref Reference temperature in K, default is td.T_REF (298.15 K)

Photosynthesis.temperature_correction — Method

temperature_correction(td::Q10{FT}, t::FT; t_ref::FT = td.T_REF) where {FT<:AbstractFloat}

Return the correction ratio for a temperature dependent variable, given

td Q10 type temperature dependency struture

t Target temperature in K

t_ref Reference temperature in K, default is td.T_REF (298.15 K)

Photosynthesis.temperature_corrected_value — Function

temperature_corrected_value(td::Union{Arrhenius{FT}, ArrheniusPeak{FT}, Q10{FT}}, t::FT; t_ref::FT = td.T_REF) where {FT<:AbstractFloat}

Return the temperature corrected value, given

td Q10 type temperature dependency struture

t Target temperature in K

t_ref Reference temperature in K, default is td.T_REF (298.15 K)

Electron transport

Photosynthesis.photosystem_electron_transport! — Function

This function updates the electron transport rates in photosynthesis reaction center, and thus to calculate photosynthetic rate. Supported methods are

photosystem_electron_transport!(psm, rc, apar)
photosystem_electron_transport!(psm, rc, apar, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/etr.jl:38.

photosystem_electron_transport!(psm, rc, apar, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/etr.jl:71.

photosystem_electron_transport!(psm, rc, apar)
photosystem_electron_transport!(psm, rc, apar, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/etr.jl:106.

Photosynthesis.photosystem_electron_transport! — Method

photosystem_electron_transport!(psm::C3VJPModel{FT}, rc::VJPReactionCenter{FT}, apar::FT, p_i::FT = FT(0)) where {FT<:AbstractFloat}

Update the electron transport rates, given

psm C3VJPModel type C3 photosynthesis model

rc VJPReactionCenter type photosynthesis system reaction center

apar Absorbed photosynthetically active radiation in μmol m⁻² s⁻¹

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthesis.photosystem_electron_transport! — Method

photosystem_electron_transport!(psm::C3CytochromeModel{FT}, rc::CytochromeReactionCenter{FT}, apar::FT, p_i::FT) where {FT<:AbstractFloat}

Update the electron transport rates, given

psm C3CytochromeModel type C4 photosynthesis model

rc CytochromeReactionCenter type photosynthesis system reaction center

apar Absorbed photosynthetically active radiation in μmol m⁻² s⁻¹

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthesis.photosystem_electron_transport! — Method

photosystem_electron_transport!(psm::C4VJPModel{FT}, rc::VJPReactionCenter{FT}, apar::FT, p_i::FT = FT(0)) where {FT<:AbstractFloat}

Update the electron transport rates, given

psm C4VJPModel type C4 photosynthesis model

rc VJPReactionCenter type photosynthesis system reaction center

apar Absorbed photosynthetically active radiation in μmol m⁻² s⁻¹

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthetic rates

Photosynthesis.rubisco_limited_rate! — Function

This function updates the RubisCO limited photosynthetic rate. Supported methods are

rubisco_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/rubisco_limited.jl:33.

rubisco_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/rubisco_limited.jl:56.

rubisco_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/rubisco_limited.jl:81.

rubisco_limited_rate!(psm, air, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/rubisco_limited.jl:107.

rubisco_limited_rate!(psm, air, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/rubisco_limited.jl:143.

Photosynthesis.rubisco_limited_rate! — Method

rubisco_limited_rate!(psm::C3VJPModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the RubisCO limited photosynthetic rate, given

psm C3VJPModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa

Photosynthesis.rubisco_limited_rate! — Method

rubisco_limited_rate!(psm::C3CytochromeModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the RubisCO limited photosynthetic rate, given

psm C3CytochromeModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa

Photosynthesis.rubisco_limited_rate! — Method

rubisco_limited_rate!(psm::C4VJPModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the RubisCO limited photosynthetic rate, given

psm C4VJPModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthesis.rubisco_limited_rate! — Method

rubisco_limited_rate!(psm::C3VJPModel{FT}, air::AirLayer{FT}, g_lc::FT) where {FT<:AbstractFloat}

Update the RubisCO limited photosynthetic rate in conductance mode, given

psm C3VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹]

Photosynthesis.rubisco_limited_rate! — Method

rubisco_limited_rate!(psm::C4VJPModel{FT}, air::AirLayer{FT}, g_lc::FT) where {FT<:AbstractFloat}

Update the RubisCO limited photosynthetic rate in conductance mode, given

psm C4VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure, not used in the method

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹], not used in this methid

Photosynthesis.light_limited_rate! — Function

This function updates the electron transport limited photosynthetic rate. Supported methods are

light_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/light_limited.jl:36.

light_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/light_limited.jl:62.

light_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/light_limited.jl:86.

light_limited_rate!(psm, air, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/light_limited.jl:111.

light_limited_rate!(psm, air, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/light_limited.jl:150.

Photosynthesis.light_limited_rate! — Method

light_limited_rate!(psm::C3VJPModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate, given

psm C3VJPModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa

Photosynthesis.light_limited_rate! — Method

light_limited_rate!(psm::C3CytochromeModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate, given

psm C3VJPModel or C3CytochromeModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa

Photosynthesis.light_limited_rate! — Method

light_limited_rate!(psm::C4VJPModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate, given

psm C4VJPModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthesis.light_limited_rate! — Method

light_limited_rate!(psm::C3VJPModel{FT}, air::AirLayer{FT}, g_lc::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate in conductance mode, given

psm C3VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹]

Photosynthesis.light_limited_rate! — Method

light_limited_rate!(psm::C4VJPModel{FT}, air::AirLayer{FT}, g_lc::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate in conductance mode, given

psm C4VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure, not used in this methid

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹], not used in this methid

Photosynthesis.product_limited_rate! — Function

This function updates the product limited photosynthetic rate. Supported methods are

product_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/product_limited.jl:33.

product_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/product_limited.jl:61.

product_limited_rate!(psm, p_i)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/product_limited.jl:86.

product_limited_rate!(psm, air, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/product_limited.jl:110.

product_limited_rate!(psm, air, g_lc)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/product_limited.jl:134.

Photosynthesis.product_limited_rate! — Method

product_limited_rate!(psm::C3VJPModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the product limited photosynthetic rate, given

psm C3VJPModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthesis.product_limited_rate! — Method

product_limited_rate!(psm::C3CytochromeModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the product limited photosynthetic rate, given

psm C3CytochromeModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa, not used in this method

Photosynthesis.product_limited_rate! — Method

product_limited_rate!(psm::C4VJPModel{FT}, p_i::FT) where {FT<:AbstractFloat}

Update the product limited photosynthetic rate, given

psm C4VJPModel structure for C3 photosynthesis model

p_i Internal CO₂ partial pressure in Pa

Photosynthesis.product_limited_rate! — Method

product_limited_rate!(psm::C3VJPModel{FT}, air::AirLayer{FT}, g_lc::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate in conductance mode, given

psm C3VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure, not used in this method

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹], not used in this method

Photosynthesis.product_limited_rate! — Method

product_limited_rate!(psm::C4VJPModel{FT}, air::AirLayer{FT}, g_lc::FT) where {FT<:AbstractFloat}

Update the electron transport limited photosynthetic rate in conductance mode, given

psm C4VJPModel structure for C3 photosynthesis model

air AirLayer structure for environmental conditions like O₂ partial pressure

g_lc Leaf diffusive conductance to CO₂ in [mol m⁻² s⁻¹]

Colimitation

Photosynthesis.colimit_photosynthesis! — Function

This function update gross and net photosynthetic rates using different colimitation method. Supported methods are

colimit_photosynthesis!(psm)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/colimit.jl:33.

colimit_photosynthesis!(psm, colim)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/colimit.jl:56.

colimit_photosynthesis!(psm, colim)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/colimit.jl:79.

colimit_photosynthesis!(psm, colim)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/colimit.jl:107.

colimit_photosynthesis!(psm, colim)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/colimit.jl:135.

Photosynthesis.colimit_photosynthesis! — Method

colimit_photosynthesis!(psm::Union{C3CytochromeModel{FT}, C3VJPModel{FT}, C4VJPModel{FT}}) where {FT<:AbstractFloat}

Use the minimal photosynthetic rates of rubisco-, light-, and product-limited photosynthetic rates, given

psm C3VJPModel or C4VJPModel type photosynthesis model

Photosynthesis.colimit_photosynthesis! — Method

colimit_photosynthesis!(psm::Union{C3VJPModel{FT}, C4VJPModel{FT}}, colim::MinimumColimit{FT}) where {FT<:AbstractFloat}

Use the minimal photosynthetic rates of rubisco-, light-, and product-limited photosynthetic rates, given

psm C3VJPModel or C4VJPModel type photosynthesis model

colim MinimumColimit type struct

Photosynthesis.colimit_photosynthesis! — Method

colimit_photosynthesis!(psm::C3CytochromeModel{FT}, colim::MinimumColimit{FT}) where {FT<:AbstractFloat}

Use the minimal photosynthetic rates of rubisco-, light-, and product-limited photosynthetic rates, given

psm C3CytochromeModel type photosynthesis model

colim MinimumColimit type struct

Photosynthesis.colimit_photosynthesis! — Method

colimit_photosynthesis!(psm::Union{C3VJPModel{FT}, C4VJPModel{FT}}, colim::QuadraticColimit{FT}) where {FT<:AbstractFloat}

Use the minimal photosynthetic rates of rubisco-, light-, and product-limited photosynthetic rates, given

psm C3VJPModel or C4VJPModel type photosynthesis model

colim QuadraticColimit type struct

Photosynthesis.colimit_photosynthesis! — Method

colimit_photosynthesis!(psm::C3CytochromeModel{FT}, colim::QuadraticColimit{FT}) where {FT<:AbstractFloat}

Use the minimal photosynthetic rates of rubisco-, light-, and product-limited photosynthetic rates, given

psm C3CytochromeModel type photosynthesis model

colim QuadraticColimit type struct

Coefficients and fluorescence

Photosynthesis.photosystem_coefficients! — Function

This function updates the rate constants and coefficients in reaction center. Supported methods are

photosystem_coefficients!(psm, rc, vdt, apar)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/fluorescence.jl:38.

photosystem_coefficients!(psm, rc, cfm, apar)

defined at /home/runner/work/Photosynthesis.jl/Photosynthesis.jl/src/fluorescence.jl:86.

Photosynthesis.photosystem_coefficients! — Method

photosystem_coefficients!(psm::Union{C3VJPModel{FT}, C4VJPModel{FT}}, rc::VJPReactionCenter{FT}, vdt::VanDerTolFluorescenceModel{FT}, apar::FT) where {FT<:AbstractFloat}

Update the rate constants and coefficients in reaction center, given

psm C3VJPModel or C4VJPModel type photosynthesis model

rc VJPReactionCenter type photosynthesis system reaction center

vdt van der Tol et al. (2013) fluorescence model

apar Absorbed photosynthetically active radiation in μmol m⁻² s⁻¹

Photosynthesis.photosystem_coefficients! — Method

photosystem_coefficients!(psm::C3CytochromeModel{FT}, rc::CytochromeReactionCenter{FT}, cfm::CytochromeFluorescenceModel{FT}, apar::FT) where {FT<:AbstractFloat}

Update the rate constants and coefficients in reaction center, given

psm C3CytochromeModel type photosynthesis model

rc CytochromeReactionCenter type photosynthesis system reaction center

cfm Johnson and Berry (2021) fluorescence model

apar Absorbed photosynthetically active radiation in μmol m⁻² s⁻¹