Photosynthesis

Temperature Dependency

Types

Photosynthesis.AbstractTDParameterSet — Type

type AbstractTDParameterSet{FT}

Hierarchy of the AbstractTDParameterSet:

ArrheniusTD

ArrheniusPeakTD

Photosynthesis.ArrheniusTD — Type

struct ArrheniusTD{FT}

An AbstractTDParameterSet type struct using

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right)\]

Fields

VAL_25

Uncorrected value at 298.15 K

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

Photosynthesis.ArrheniusPeakTD — Type

struct ArrheniusPeakTD{FT}

An AbstractTDParameterSet type struct using

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

Parameter sets

Photosynthesis.JmaxTDBernacchi — Function

ArrheniusPeakTD type Jmax TD

Photosynthesis.JmaxTDCLM — Function

ArrheniusPeakTD type Jmax TD

Photosynthesis.JmaxTDLeuning — Function

ArrheniusPeakTD type Jmax TD from Leuning's data

Photosynthesis.KcTDBernacchi — Function

ArrheniusTD type Kc TD from Bernacchi's data

Photosynthesis.KcTDCLM — Function

ArrheniusTD type Kc TD

Photosynthesis.KoTDBernacchi — Function

ArrheniusTD type Ko TD from Bernacchi's data

Photosynthesis.KoTDCLM — Function

ArrheniusTD type Ko TD

Photosynthesis.KpepTDBoyd — Function

ArrheniusTD type Kpep TD from Boyd's data

Photosynthesis.KpepTDCLM — Function

ArrheniusTD type Kpep TD

Photosynthesis.RespirationTDBernacchi — Function

ArrheniusTD type Respiration TD from Bernacchi's data

Photosynthesis.RespirationTDCLM — Function

ArrheniusPeakTD type Respiration TD

Photosynthesis.VcmaxTDBernacchi — Function

ArrheniusTD type Vcmax TD from Bernacchi's data

Photosynthesis.VcmaxTDCLM — Function

ArrheniusPeakTD type Vcmax TD

Photosynthesis.VcmaxTDLeuning — Function

ArrheniusPeakTD type Vcmax TD from Leuning's data

Photosynthesis.VomaxTDBernacchi — Function

ArrheniusTD type Vomax TD from Bernacchi's data

Photosynthesis.VpmaxTDBoyd — Function

ArrheniusPeakTD type Vpmax TD from Boyd's data

Photosynthesis.ΓStarTDBernacchi — Function

ArrheniusTD type $Γ^{*}$ TD from Bernacchi's data

Photosynthesis.ΓStarTDCLM — Function

ArrheniusTD type Γ* TD

Fluorescence Model

Types

Photosynthesis.AbstractFluoModelParaSet — Type

type AbstractFluoModelParaSet

Hierarchy of the AbstractFluoModelParaSet:

FluoParaSet

Photosynthesis.FluoParaSet — Type

struct FluoParaSet{FT}

A AbstractFluoModelParaSet type paramter set.

Fields

Kn1

Fluorescence model coefficient

Kn2

Fluorescence model coefficient

Kn3

Fluorescence model coefficient

Parameter sets

Photosynthesis.FluorescenceFlexas — Function

FluoParaSet type parameter set using Flexas's data

Respiration Quantification

Photosynthesis.VtoRCollatz — Function

A constant of 0.01

Photosynthesis.VtoRDefault — Function

A constant of 0.015

Parameter Sets Collection

Types

Photosynthesis.AbstractPhotoModelParaSet — Type

type AbstractPhotoModelParaSet

Hierarchy of the AbstractPhotoModelParaSet:

C3ParaSet

C4ParaSet

Photosynthesis.C3ParaSet — Type

struct C3Paraset{FT}

Parameter sets for C3 photosynthesis.

Fields

JT

Jmax temperature dependency

KcT

Kc temperature dependency

KoT

Ko temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

ΓsT

Γ_star temperature dependency

Flu

Fluorescence model

VR

Vcmax25 and respiration correlation

Eff_1

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

Eff_2

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

Photosynthesis.C4ParaSet — Type

struct C4ParaSet{FT}

Parameter sets for C3 photosynthesis.

Fields

KpT

Kpep temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

VpT

Vpmax temperature dependency

Flu

Fluorescence model

VR

Vcmax25 and respiration correlation

Parameter sets

Photosynthesis.C3Bernacchi — Function

C3ParaSet type C3 photosynthesis using Bernacchi's data

Photosynthesis.C3CLM — Function

C3ParaSet type C3 photosynthesis using CLM5's data

Photosynthesis.C4CLM — Function

C4ParaSet type C4 photosynthesis using CLM5's data

Leaf and Environment

Photosynthesis.Leaf — Type

struct Leaf{FT}

Struct to store leaf information.

Fields

T

Temperature [K]

T_old

Temperature history [K]

Kd

Rate constant for thermal dissipation

Kf

Rate constant for fluorescence (const)

Kn

NPQ rate constant (initially zero)

Kp

Rate constant for photochemistry (all reaction centers open)

maxPSII

max PSII yield (Kn=0, all RC open)

PSII_frac

Fraction of absorbed light used by PSII ETR

p_i

Leaf internal CO₂ partial pressure [Pa]

g_lc

Leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

Ac

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

Aj

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

Ag

Gross photosynthetic rate [μmol m⁻² s⁻¹]

An

Net photosynthetic rate [μmol m⁻² s⁻¹]

Ap

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

J

Electron transport [μmol m⁻² s⁻¹]

J_pot

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

Jmax

Maximal electron transport rate [μmol m⁻² s⁻¹]

Jmax25

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

Kc

RubisCO coefficient Kc [Pa]

Ko

RubisCO coefficient Ko [Pa]

Kpep

PEP coefficient Ko [Pa]

Km

Michaelis-Menten's coefficient [Pa]

Rd

Respiration rate [μmol m⁻² s⁻¹]

Rd25

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

Vcmax

Maximal carboxylation rate [μmol m⁻² s⁻¹]

Vcmax25

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

Vpmax

Maximal PEP carboxylation rate [μmol m⁻² s⁻¹]

Vpmax25

Maximal PEP carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

Γ_star

CO₂ compensation point with the absence of Rd [Pa]

CO₂_per_electron

Total efficiency, incl. photorespiration [mol CO₂ mol⁻¹ e-]

Fm

dark adapted yield (Kp=0)

Fm′

light adapted yield (Kp=0)

Fo

dark-adapted fluorescence yield (Kp=max)

Fo′

light-adapted fluorescence yield in the dark (Kp=max)

Ja

Actual electron transport rate [μmol m⁻² s⁻¹]

NPQ

Non-Photochemical quenching

qQ

Photochemical quenching

qE

energy quenching

φ

PSII yield

ϕs

Steady-state (light-adapted) yield (aka Fs)

APAR

Absorbed photosynthetic active radiation [μmol m⁻² s⁻¹]

Photosynthesis.AirLayer — Type

struct AirLayer{FT}

Struct to store environmental conditions in each air layer corresponds to one canopy layer.

Fields

t_air

Air temperature [K]

p_a

Atmospheric CO₂ partial pressure [Pa]

p_atm

Atmospheric pressure [Pa]

p_H₂O

Atmospheric vapor pressure [Pa]

p_O₂

Atmospheric O₂ partial pressure [Pa]

p_sat

Saturation vapor pressure [Pa]

RH

Relatiev humidity

Photosynthesis Model

Temperature corrections

Photosynthesis.arrhenius_correction — Function

arrhenius_correction(td_set::AbstractTDParameterSet{FT}, T::FT)

A correction factor based on arrhenius's fitting procedure, given

td_set ArrheniusTD or ArrheniusPeakTD type struct

T Leaf temperature in [K]

The equation used for ArrheniusTD is

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right)\]

The equations used for ArrheniusPeakTD are

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Photosynthesis.photo_TD_from_set — Function

photo_TD_from_set(td_set::ArrheniusTD{FT}, T::FT)

Make temperature correction from parameter set, given

td_set ArrheniusTD type parameter set, which has a VAL_25 field

T Leaf temperature

Useful for Kc, Ko, Kpep, and $Γ^{*}$.

Photosynthesis.photo_TD_from_val — Function

photo_TD_from_val(td_set::AbstractTDParameterSet, val::FT, T::FT)

Make temperature correction from a given value, given

td_set ArrheniusTD or ArrheniusPeakTD type struct

val Uncorrected value at 298.15 K

T Leaf temperature

Useful for Vcmax, Vomac, Vpmax, Jmax, and Respiration.

Photosynthesis.leaf_jmax! — Function

leaf_jmax!(td_set::AbstractTDParameterSet, leaf::Leaf{FT})

Update maximal electron transport rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_kc! — Function

leaf_kc!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update Kc at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_km! — Function

leaf_km!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT})

Update Ko at leaf temperature, given

photo_set C3ParaSet type photosynthesis parameter set

leaf Leaf type struct

envir AirLayer type struct

Photosynthesis.leaf_ko! — Function

leaf_ko!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update Ko at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_kpep! — Function

leaf_kpep!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update Kpep at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_rd! — Function

leaf_rd!(td_set::AbstractTDParameterSet, leaf::Leaf)

Update leaf dark respiration rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_vcmax! — Function

leaf_vcmax!(td_set::AbstractTDParameterSet, leaf::Leaf)

Update leaf maximal carboxylation rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_vpmax! — Function

leaf_vpmax!(td_set::AbstractTDParameterSet, leaf::Leaf)

Update leaf maximal PEP carboxylation rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_Γstar! — Function

leaf_Γstar!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT})

Update $Γ^{*}$ at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthetic rates (individual)

Photosynthesis.rubisco_limited_rate! — Function

rubisco_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
rubisco_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Calculate the RubisCO limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Photosynthesis.rubisco_limited_rate_glc! — Function

rubisco_limited_rate_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Calculate the RubisCO limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Photosynthesis.leaf_ETR! — Function

leaf_ETR!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
leaf_ETR!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Update the electron transport variables in the leaf struct, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Photosynthesis.light_limited_rate! — Function

light_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
light_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Calculate the Light limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Photosynthesis.light_limited_rate_glc! — Function

light_limited_rate_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Calculate the Light limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Photosynthesis.product_limited_rate! — Function

product_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT})
product_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT})

Calculate the Product limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Photosynthesis.product_limited_rate_glc! — Function

product_limited_rate_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Calculate the Product limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Photosynthesis.leaf_fluorescence! — Function

leaf_fluorescence!(fluo_set::FluoParaSet{FT}, leaf::Leaf{FT})

Compute fluorescence yield, Kn, and Kp for leaf, given

fluo_set FluoParaSet type parameter set

leaf Leaf struct

Photosynthetic rates

Photosynthesis.leaf_temperature_dependence! — Function

leaf_temperature_dependence!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::AbstractLeaf,
        envir::AirLayer{FT})

Update the temperature dependent photosynthesis only, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

Photosynthesis.leaf_photo_from_pi! — Function

leaf_photo_from_pi!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::AbstractLeaf,
        envir::AirLayer{FT})

Compute leaf photosynthetic rates, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

The C3 photosynthesis model is from Farquhar et al. (1980) "A biochemical model of photosynthetic CO₂ assimilation in leaves of C3 species."

The C4 photosynthesis model is adapted from Collatz et al. (1992) "Coupled photosynthesis-stomatal conductance model for leaves of C4 plants."

Photosynthesis.leaf_photo_from_glc! — Function

leaf_photo_from_glc!(
        photo_set::AbstractPhotoModelParaSet,
        leaf::Leaf{FT},
        envir::AirLayer{FT})

Update leaf photosynthetic rates from a known leaf diffusive conductance, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct