Photosynthesis

Leaf and Environment Structures

To model photosynthesis more efficiently, we use a container (Leaf struct) to store the photosynthesis-related information. For example, many of the physiological parameters are temperature-dependent, but these temperature-dependent values only need to be updated when leaf temperature changes. Therefore, use of the container significantly reduces the time required when programing leaf gas exchange prognostically. The Leaf struct has the following fields:

Photosynthesis.Leaf — Type

mutable struct Leaf{FT}

Struct to store leaf information.

Fields

T

Temperature [K]

Kd

Rate constant for thermal dissipation

Kf

Rate constant for fluorescence (const)

Kn

NPQ rate constant (initially zero)

Kp

Rate constant for photochemistry (all reaction centers open)

maxPSII

max PSII yield (Kn=0, all RC open)

PSII_frac

Fraction of absorbed light used by PSII ETR

p_i

Leaf internal CO₂ partial pressure [Pa]

p_s

Leaf surface CO₂ partial pressure [Pa]

p_sat

Saturation H₂O vapor pressure [Pa]

g_bc

Leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

g_lc

Leaf diffusive conductance to CO₂ [mol m⁻² s⁻¹]

Ac

RubisCO limited photosynthetic rate [μmol m⁻² s⁻¹]

Aj

Light limited photosynthetic rate [μmol m⁻² s⁻¹]

Ag

Gross photosynthetic rate [μmol m⁻² s⁻¹]

An

Net photosynthetic rate [μmol m⁻² s⁻¹]

Ap

Product limited photosynthetic rate [μmol m⁻² s⁻¹]

J

Electron transport [μmol m⁻² s⁻¹]

J_pot

Potential Electron Transport Rate [μmol m⁻² s⁻¹]

Jmax

Maximal electron transport rate [μmol m⁻² s⁻¹]

Jmax25

Maximal electron transport rate at 298.15 K [μmol m⁻² s⁻¹]

Kc

RubisCO coefficient Kc [Pa]

Ko

RubisCO coefficient Ko [Pa]

Kpep

PEP coefficient Ko [Pa]

Km

Michaelis-Menten's coefficient [Pa]

Rd

Respiration rate [μmol m⁻² s⁻¹]

Rd25

Respiration rate at 298.15 K [μmol m⁻² s⁻¹]

Vcmax

Maximal carboxylation rate [μmol m⁻² s⁻¹]

Vcmax25

Maximal carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

Vpmax

Maximal PEP carboxylation rate [μmol m⁻² s⁻¹]

Vpmax25

Maximal PEP carboxylation rate at 298.15 K [μmol m⁻² s⁻¹]

Γ_star

CO₂ compensation point with the absence of Rd [Pa]

CO₂_per_electron

Total efficiency, incl. photorespiration [mol CO₂ mol⁻¹ e-]

Fm

dark adapted yield (Kp=0)

Fm′

light adapted yield (Kp=0)

Fo

dark-adapted fluorescence yield (Kp=max)

Fo′

light-adapted fluorescence yield in the dark (Kp=max)

Ja

Actual electron transport rate [μmol m⁻² s⁻¹]

NPQ

Non-Photochemical quenching

qQ

Photochemical quenching

qE

energy quenching

φ

PSII yield

ϕs

Steady-state (light-adapted) yield (aka Fs)

APAR

Absorbed photosynthetic active radiation [μmol m⁻² s⁻¹]

Also, environmental conditions are required to compute photosynthetic rate, and these conditions are stored in AirLayer struct. An AirLayer struct further allows for more conveniently modeling photosynthesis the vertical CO₂ and H₂O gradients in the canopy. The AirLayer structs has the following fields:

Photosynthesis.AirLayer — Type

mutable struct AirLayer{FT}

Struct to store environmental conditions in each air layer corresponds to one canopy layer.

Fields

t_air

Air temperature [K]

p_a

Atmospheric CO₂ partial pressure [Pa]

p_atm

Atmospheric pressure [Pa]

p_H₂O

Atmospheric vapor pressure [Pa]

p_O₂

Atmospheric O₂ partial pressure [Pa]

p_sat

Saturation vapor pressure [Pa]

RH

Relatiev humidity

See exmaples below for how to create the structs

using Photosynthesis

FT = Float32;
leaf = Leaf{FT}();
envir = AirLayer{FT}();

Temperature Dependency Structs

The temperature-dependent (TD) photosynthetic parameters include

$J_\text{max}$ Maximal electron transport rate

$K_\text{c}$ Michaelis constant for CO₂

$K_\text{m}$ Michaelis-Menten's coefficient

$K_\text{o}$ Michaelis constant for O₂

$K_\text{pep}$ Michaelis constant for PEP carboxylation

$R_\text{d}$ Dark respiration

$V_\text{cmax}$ Maximal RuBP carboxylation rate

$V_\text{omax}$ Maximal RuBP oxygenation rate

$V_\text{pmax}$ Maximal PEP carboxylation rate

$Γ^{*}$ CO₂ compensation point with the absence of dark respiration

There are two typical types of temperature dependencies using the classic Arrhenius equation. We define the two types as ArrheniusTD and ArrheniusPeakTD subject to AbstractTDParameterSet type:

Photosynthesis.AbstractTDParameterSet — Type

abstract type AbstractTDParameterSet{FT}

Hierarchy of the AbstractTDParameterSet:

ArrheniusTD

ArrheniusPeakTD

Photosynthesis.ArrheniusTD — Type

struct ArrheniusTD{FT}

An AbstractTDParameterSet type struct using

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right)\]

Fields

VAL_25

Uncorrected value at 298.15 K

ΔHa_to_R

Ratio between ΔHa and R [K]

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

Photosynthesis.ArrheniusPeakTD — Type

struct ArrheniusPeakTD{FT}

An AbstractTDParameterSet type struct using

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Fields

ΔHa_to_RT25

Ratio between ΔHa and R*K_25

ΔHd_to_R

Ratio between ΔHd and R

ΔSv_to_R

Ratio between ΔSv and R

C

Correction factor C = 1 + exp( Sv/R + Hd/(RT0) )

There are many published parameter sets for the various temperature dependencies, and to ease the modeling we predefined most of the structs:

Photosynthesis.JmaxTDBernacchi — Function

ArrheniusPeakTD type Jmax TD

Photosynthesis.JmaxTDCLM — Function

ArrheniusPeakTD type Jmax TD

Photosynthesis.JmaxTDLeuning — Function

ArrheniusPeakTD type Jmax TD from Leuning's data

Photosynthesis.KcTDBernacchi — Function

ArrheniusTD type Kc TD from Bernacchi's data

Photosynthesis.KcTDCLM — Function

ArrheniusTD type Kc TD

Photosynthesis.KoTDBernacchi — Function

ArrheniusTD type Ko TD from Bernacchi's data

Photosynthesis.KoTDCLM — Function

ArrheniusTD type Ko TD

Photosynthesis.KpepTDBoyd — Function

ArrheniusTD type Kpep TD from Boyd's data

Photosynthesis.KpepTDCLM — Function

ArrheniusTD type Kpep TD

Photosynthesis.RespirationTDBernacchi — Function

ArrheniusTD type Respiration TD from Bernacchi's data

Photosynthesis.RespirationTDCLM — Function

ArrheniusPeakTD type Respiration TD

Photosynthesis.VcmaxTDBernacchi — Function

ArrheniusTD type Vcmax TD from Bernacchi's data

Photosynthesis.VcmaxTDCLM — Function

ArrheniusPeakTD type Vcmax TD

Photosynthesis.VcmaxTDLeuning — Function

ArrheniusPeakTD type Vcmax TD from Leuning's data

Photosynthesis.VomaxTDBernacchi — Function

ArrheniusTD type Vomax TD from Bernacchi's data

Photosynthesis.VpmaxTDBoyd — Function

ArrheniusPeakTD type Vpmax TD from Boyd's data

Photosynthesis.ΓStarTDBernacchi — Function

ArrheniusTD type $Γ^{*}$ TD from Bernacchi's data

Photosynthesis.ΓStarTDCLM — Function

ArrheniusTD type Γ* TD

The TDs can be easily created using commands like

using Photosynthesis

FT = Float32;
_td_1 = JmaxTDBernacchi(FT);
_td_2 = VcmaxTDCLM(FT);

However, be aware that these pre-defined TD structs are not mutable, to create customized TD struct, code like this will be useful

using Photosynthesis

FT = Float32;
_td_1 = ArrheniusTD{FT}(1, 10000, 30);
_td_1 = ArrheniusPeakTD{FT}(1, 10000, 30, 1);

To further simplify the use of Photosynthesis module, we provide a few collections/structs of temperature dependencies as well as other parameter sets like FluoParaSet. The structs are catergorized to C3ParaSet and C4ParaSet subject to an AbstractPhotoModelParaSet type, and the structs are meant for modeling C3 photosynthesis and C4 photosynthesis, respectively.

Photosynthesis.AbstractPhotoModelParaSet — Type

abstract type AbstractPhotoModelParaSet{FT}

Hierarchy of the AbstractPhotoModelParaSet:

C3ParaSet

C4ParaSet

Photosynthesis.C3ParaSet — Type

mutable struct C3Paraset{FT}

Parameter sets for C3 photosynthesis.

Fields

JT

Jmax temperature dependency

KcT

Kc temperature dependency

KoT

Ko temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

ΓsT

Γ_star temperature dependency

Flu

Fluorescence model

VR

Vcmax25 and respiration correlation

Eff_1

Coefficient 4.0/4.5 for NADPH/ATP requirement stochiometry, respectively

Eff_2

Coefficient 8.0/10.5 for NADPH/ATP requirement stochiometry, respectively

Photosynthesis.C4ParaSet — Type

mutable struct C4ParaSet{FT}

Parameter sets for C3 photosynthesis.

Fields

KpT

Kpep temperature dependency

ReT

Respiration temperature dependency

VcT

Vcmax temperature dependency

VpT

Vpmax temperature dependency

Flu

Fluorescence model

VR

Vcmax25 and respiration correlation

Again, to guarantee a quick start, we provided a few pre-defined parameter sets:

Photosynthesis.C3Bernacchi — Function

C3ParaSet type C3 photosynthesis using Bernacchi's data

Photosynthesis.C3CLM — Function

C3ParaSet type C3 photosynthesis using CLM5's data

Photosynthesis.C4CLM — Function

C4ParaSet type C4 photosynthesis using CLM5's data

Examples:

using Photosynthesis

FT = Float32;
set_b = C3Bernacchi(FT);
set_3 = C3CLM(FT);
set_4 = C4CLM(FT);

Note it here that the C3ParaSet and C4ParaSet structs are mutable, and the fields can be changed to another non-mutable TD struct. We'd like to mention that in some cases, leaf respiration rate is not measured, and in this case, the dark respiration rate will be computed from $V_\text{cmax}$ using a multiplier

Photosynthesis.VtoRCollatz — Function

A constant of 0.01

Photosynthesis.VtoRDefault — Function

A constant of 0.015

Temperature Dependency

As mentioned above, temperature corrections only need to be done once per temperature change, and storing the temperature corrected values will significantly boost the code speed. Here we provide a few functions to change the stored values. First of all, all the temperature corrections are made with arrhenius_correction:

Photosynthesis.arrhenius_correction — Function

arrhenius_correction(td_set::ArrheniusTD{FT}, T::FT) where {FT<:AbstractFloat}
arrhenius_correction(td_set::ArrheniusPeakTD{FT}, T::FT) where {FT<:AbstractFloat}

A correction factor based on arrhenius's fitting procedure, given

td_set ArrheniusTD or ArrheniusPeakTD type struct

T Leaf temperature in [K]

The equation used for ArrheniusTD is

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right)\]

The equations used for ArrheniusPeakTD are

\[corr = \exp \left( \dfrac{ΔHa}{R T_0} - \dfrac{ΔHa}{R T_1} \right) \cdot \dfrac{ 1 + \exp \left( \dfrac{S_v T_0 - H_d}{R T_0} \right) } { 1 + \exp \left( \dfrac{S_v T_1 - H_d}{R T_1} \right) }\]

Second, depending on which physiological parameter to correct, some corrections use the VAL_25 field in the ArrheniusTD, like $K_\text{c}$, $K_\text{o}$, and $K_\text{pep}$:

Photosynthesis.photo_TD_from_set — Function

photo_TD_from_set(td_set::ArrheniusTD{FT}, T::FT) where {FT<:AbstractFloat}

Make temperature correction from parameter set, given

td_set ArrheniusTD type parameter set, which has a VAL_25 field

T Leaf temperature

Useful for Kc, Ko, Kpep, and $Γ^{*}$.

Some corrections use the reference values from the Leaf struct, like $V_\text{cmax}$ and $J_\text{max}$:

Photosynthesis.photo_TD_from_val — Function

photo_TD_from_val(td_set::AbstractTDParameterSet{FT}, val::FT, T::FT) where {FT<:AbstractFloat}

Make temperature correction from a given value, given

td_set ArrheniusTD or ArrheniusPeakTD type struct

val Uncorrected value at 298.15 K

T Leaf temperature

Useful for Vcmax, Vomac, Vpmax, Jmax, and Respiration.

The functions to make temperature corrections to each individual variables are

Photosynthesis.leaf_jmax! — Function

leaf_jmax!(td_set::AbstractTDParameterSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update maximal electron transport rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_kc! — Function

leaf_kc!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update Kc at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_km! — Function

leaf_km!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Update Ko at leaf temperature, given

photo_set C3ParaSet type photosynthesis parameter set

leaf Leaf type struct

envir AirLayer type struct

Photosynthesis.leaf_ko! — Function

leaf_ko!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update Ko at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_kpep! — Function

leaf_kpep!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update Kpep at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_rd! — Function

leaf_rd!(td_set::AbstractTDParameterSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update leaf dark respiration rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_vcmax! — Function

leaf_vcmax!(td_set::AbstractTDParameterSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update leaf maximal carboxylation rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_vpmax! — Function

leaf_vpmax!(td_set::AbstractTDParameterSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update leaf maximal PEP carboxylation rate at leaf temperature, given

td_set AbstractTDParameterSet type TD parameter set

leaf Leaf type struct

Photosynthesis.leaf_Γstar! — Function

leaf_Γstar!(td_set::ArrheniusTD{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update $Γ^{*}$ at leaf temperature, given

td_set ArrheniusTD type TD parameter set

leaf Leaf type struct

Again to ease the coding, we provide a function to run all the temperature dependencies:

Photosynthesis.leaf_temperature_dependence! — Function

leaf_temperature_dependence!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}
leaf_temperature_dependence!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}
leaf_temperature_dependence!(photo_set::AbstractPhotoModelParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, T::FT) where {FT<:AbstractFloat}

Update the temperature dependent photosynthesis only, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

T Given leaf temperature

Note it here that function leaf_temperature_dependence! updates saturated vapor pressure from leaf temperature as well.

Example:

using Photosynthesis

FT = Float32;
leaf = Leaf{FT}();
envir = AirLayer{FT}();
set_3 = C3CLM(FT);

leaf_temperature_dependence!(c3_set, leaf, envir);
leaf_temperature_dependence!(c3_set, leaf, envir, FT(300));

RubisCO-limited Photosynthesis

By default, Photosynthesis module computes gross photosynthetic rate as the minimal of the three:

$A_\text{c}$ RubisCO-limited photosynthetic rate

$A_\text{j}$ Light-limited photosynthetic rate

$A_\text{p}$ Product-limited photosynthetic rate

If leaf internal CO₂ is known, $A_\text{c}$ (gross rate) can be computed using

Photosynthesis.rubisco_limited_rate! — Function

rubisco_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}
rubisco_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Calculate the RubisCO limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

If total leaf diffusive conductance to CO₂ is known, $A_\text{c}$ can be computed analytically using

Photosynthesis.rubisco_limited_rate_glc! — Function

rubisco_limited_rate_glc!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Calculate the RubisCO limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Note it here that rubisco_limited_rate_glc! only applies to C3 photosynthesis as the RubisCO-limited rate for C4 plants is $V_\text{cmax}$.

To faciliate the computation of photosynthetic rate for a number of leaves with different conductances but same temperature (as in the case of CliMA Land model), function rubisco_limited_an_glc returns (an array of) net photosynthetic rate limited by RubisCO.

Photosynthesis.rubisco_limited_an_glc — Function

rubisco_limited_an_glc(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, glcs::Array{FT,1}) where {FT<:AbstractFloat}
rubisco_limited_an_glc(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, glcs::Array{FT,1}) where {FT<:AbstractFloat}

Calculate the RubisCO limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

glcs Given array of g_lc

Light-limited Photosynthesis

If leaf internal CO₂ is known, $A_\text{j}$ (gross rate) can be computed using

Photosynthesis.light_limited_rate! — Function

light_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}
light_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Calculate the Light limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

If total leaf diffusive conductance to CO₂ is known, $A_\text{j}$ can be computed analytically using

Photosynthesis.light_limited_rate_glc! — Function

light_limited_rate_glc!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Calculate the Light limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Note it here that light_limited_rate_glc! only applies to C3 photosynthesis as the RubisCO-limited rate for C4 plants is the electron transport rate.

To faciliate the computation of photosynthetic rate for a number of leaves with different conductances but same temperature (as in the case of CliMA Land model), function light_limited_an_glc returns (an array of) net photosynthetic rate limited by light.

Photosynthesis.light_limited_an_glc — Function

light_limited_an_glc(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, glcs::Array{FT,1}, Js::Array{FT,1}) where {FT<:AbstractFloat}
light_limited_an_glc(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, glcs::Array{FT,1}, Js::Array{FT,1}) where {FT<:AbstractFloat}

Calculate the Light limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

glcs Given array of g_lc

Js Given array of electron transport rate

Be aware that leaf electron transport rate needs to be calculated before the light-limited rate:

Photosynthesis.leaf_ETR! — Function

leaf_ETR!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}
leaf_ETR!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Update the electron transport variables in the leaf struct, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Photosynthesis.leaf_ETR_Jps — Function

leaf_ETR_Jps(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, Jps::Array{FT,1}) where {FT<:AbstractFloat}
leaf_ETR_Jps(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, Jps::Array{FT,1}) where {FT<:AbstractFloat}

Update the electron transport variables in the leaf struct, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

Jps Given Array of potential ETR

Photosynthesis.leaf_ETR_pot_APAR — Function

leaf_ETR_pot_APAR(leaf::Leaf{FT}, PARs::Array{FT,1}) where {FT<:AbstractFloat}

Update the electron transport variables in the leaf struct, given

leaf Leaf type struct

PARs Given Array of PAR

Product-limited Photosynthesis

If leaf internal CO₂ is known, $A_\text{p}$ (gross rate) can be computed using

Photosynthesis.product_limited_rate! — Function

product_limited_rate!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}
product_limited_rate!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Calculate the Product limited photosynthetic rate, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

If total leaf diffusive conductance to CO₂ is known, $A_\text{p}$ can be computed analytically using

Photosynthesis.product_limited_rate_glc! — Function

product_limited_rate_glc!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}

Calculate the Product limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

Note it here that product_limited_rate_glc! only applies to C4 photosynthesis as the RubisCO-limited rate for C4 plants is $V_\text{cmax}$/2.

To faciliate the computation of photosynthetic rate for a number of leaves with different conductances but same temperature (as in the case of CliMA Land model), function product_limited_an_glc returns (an array of) product-limited net photosynthetic rate.

Photosynthesis.product_limited_an_glc — Function

product_limited_an_glc(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, glcs::Array{FT,1}) where {FT<:AbstractFloat}
product_limited_an_glc(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, glcs::Array{FT,1}) where {FT<:AbstractFloat}

Calculate the Product limited photosynthetic rate from glc, given

photo_set C3ParaSet or C4ParaSet type struct

leaf Leaf type struct

envir AirLayer type struct

glcs Given array of g_lc

Photosynthetic Rates

It is worth noting that rubisco_limited_an_glc, light_limited_an_glc, and product_limited_an_glc are more efficient than computing the rates for each individual leaf. However, for empirical and optimization stomatal models, iterations are required to get the final solution as in StomataModels module. In this case, more conveniently computing photosynthetic rates for each leaf is preferable. In this case, leaf_photo_from_pi! and leaf_photo_from_glc! are better options:

Photosynthesis.leaf_photo_from_pi! — Function

leaf_photo_from_pi!(photo_set::AbstractPhotoModelParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}
leaf_photo_from_pi!(photo_set::AbstractPhotoModelParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}, p_i::FT) where {FT<:AbstractFloat}

Compute leaf photosynthetic rates, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

p_i Given leaf internal CO₂

The C3 photosynthesis model is from Farquhar et al. (1980) "A biochemical model of photosynthetic CO₂ assimilation in leaves of C3 species."

The C4 photosynthesis model is adapted from Collatz et al. (1992) "Coupled photosynthesis-stomatal conductance model for leaves of C4 plants."

Photosynthesis.leaf_photo_from_glc! — Function

leaf_photo_from_glc!(photo_set::C3ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}
leaf_photo_from_glc!(photo_set::C4ParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT}) where {FT<:AbstractFloat}
leaf_photo_from_glc!(photo_set::AbstractPhotoModelParaSet{FT}, leaf::Leaf{FT}, envir::AirLayer{FT},g_lc::FT) where {FT<:AbstractFloat}

Update leaf photosynthetic rates from a known leaf diffusive conductance, given

photo_set AbstractPhotoModelParaSet type parameter set

leaf Leaf type struct

envir AirLayer type struct

g_lc Given leaf diffusive conductance to CO₂

Fluorescence

Photosynthesis module also provide ways to compute leaf fluorescence. By default, the modules uses fluorescence parameters from Flexas et al. with struct FluorescenceFlexas:

Photosynthesis.AbstractFluoModelParaSet — Type

abstract type AbstractFluoModelParaSet{FT}

Hierarchy of the AbstractFluoModelParaSet:

FluoParaSet

Photosynthesis.FluoParaSet — Type

mutable struct FluoParaSet{FT}

A AbstractFluoModelParaSet type paramter set.

Fields

Kn1

Fluorescence model coefficient

Kn2

Fluorescence model coefficient

Kn3

Fluorescence model coefficient

Photosynthesis.FluorescenceFlexas — Function

FluoParaSet type parameter set using Flexas's data

The function that is used to compute fluorescene is

Photosynthesis.leaf_fluorescence! — Function

leaf_fluorescence!(fluo_set::FluoParaSet{FT}, leaf::Leaf{FT}) where {FT<:AbstractFloat}

Compute fluorescence yield, Kn, and Kp for leaf, given

fluo_set FluoParaSet type parameter set

leaf Leaf struct